Possible Fix for ESP dupe bug
So I've studied the bug, and it seems that it only occurs when other
texts other than ESP show up, like the intro and tab menu. It also only
happens to one ESP entry, even if it doesn't have a text table to dupe.
So what I've done is brought back the ability to see your own ESP entry.
and removed the extra info plugin call if the entry is yours. It seems
to have fixed it so far.