Fixed a mysql error related to changed uniqueID's
When a uniqueID has changed, ItemFindByID still sometimes finds the item
(or another one) due to partial matching. This means the uniqueID would
be different from k and this would create an error when adding the
instance: the table was made using uniqueID, but the item was added with
a not necessarily same 'k'.
Updated the code to avoid errors, and just not re-add items if the
correct one wasn't found (to avoid people suddenly gaining items they
really shouldn't have).