Changed CanCraft to check if OnCraft returns true
I thought that a custom check would be useful for crafting recipes involving custom systems. You could use OnCraft to check for a type of entity near the player for the recipe or, if you're spawning an entity, check to see if there is room in the direction you're facing.