Offload state changing logic from query function
Ideally, the query function should simply return a decision based on the state - it should not influence the state itself. Moving the code responsible for setting the property table to nil from the PlayerCanTakeFromStorage query function to the PlayerTakeFromStorage function means that two successive calls to the query function will return the same result.