Fix invisible characterModel
When the panel to the left of the characterModel is too small (due to only one row of models for the player to select from) the y will be calculated and set to a negative position. This means that the characterModel will not render at all. Therefore, we clamp the y to be greater than zero to ensure that this doesn't happen.